added support for Spring boot admin and chatbot

+++ src/main/java/com/example/webBanXeApi/service/NewsService.java
@@ -23,7 +23,7 @@ public class NewsService {
         return pagedResult.toList();
     }
 
-    public News getNewsById(Integer id) {
+    public News getNewsById(Long id) {
         return newsRepository.findById(id).orElseThrow(() -> new RuntimeException("News not found"));
     }
 
@@ -31,7 +31,7 @@ public class NewsService {
         return newsRepository.save(news);
     }
 
-    public News updateNews(Integer id, News news) {
+    public News updateNews(Long id, News news) {
         News existingNews = getNewsById(id);
         existingNews.setTitle(news.getTitle());
         existingNews.setContent(news.getContent());
@@ -42,7 +42,7 @@ public class NewsService {
         return newsRepository.save(existingNews);
     }
 
-    public void deleteNews(Integer id) {
+    public void deleteNews(Long id) {
         newsRepository.deleteById(id);
     }
